agetty: move network code to issuefile-netlink.c
Create issuefile-netlink.c with:
- network helper functions (print_iface_best, print_addrq_bestofall,
  dump_iface_good, dump_iface_all) moved from issuefile.c
- handler functions for \4, \6, \a, \A escapes
- netlink initialization (conditional on parsed ifile content)
- agetty_issue_register_netlink() to set up everything

The registration function checks whether the parsed issue file
contains any network escapes before initializing netlink, avoiding
unnecessary setup when no network information is requested.

Remove output_special_char(), get_escape_argument(), and the inline
netlink setup block from issuefile.c — all escape handling is now
done through the parser handler infrastructure.
