+++ agetty-cmd/Makemodule.am
@@ -7,6 +7,7 @@ agetty_SOURCES = agetty-cmd/agetty.c \
 		 agetty-cmd/agetty.h \
 		 agetty-cmd/credentials.c \
 		 agetty-cmd/issuefile.c \
+		 agetty-cmd/issuefile-netlink.c \
 		 agetty-cmd/issuefile-parser.c \
 		 agetty-cmd/issuefile-parser.h \
 		 agetty-cmd/tty.c \
