+++ agetty-cmd/issuefile.c
@@ -36,11 +36,6 @@
 # include <systemd/sd-login.h>
 #endif
 
-#ifdef USE_NETLINK
-# include <net/if.h>
-# include <arpa/inet.h>
-static uint32_t netlink_groups;
-#endif
 
 static char *read_os_release(struct agetty_options *op, const char *varname)
 {
@@ -359,6 +354,8 @@ static void register_handlers(struct agetty_issue *ie)
 	agetty_ifile_set_handler(ls, AGETTY_ESC_USERS_TEXT, print_users, NULL, NULL);
 
 	agetty_ifile_set_handler(ls, AGETTY_ESC_ESCAPE, print_escape_char, NULL, NULL);
+
+	agetty_issue_register_netlink(ie);
 }
 
 #ifdef AGETTY_RELOAD
@@ -466,42 +463,6 @@ void agetty_issue_eval(struct agetty_issue *ie,
 	if (agetty_ifile_is_empty(&ie->ifile))
 		goto done;
 
-#ifdef USE_NETLINK
-/* TODO:
- * Two pass processing for agetty_issue_eval()
- * Implement pass 1: Just evaluate list of netlink_groups (IP protocols) and
- * interfaces to monitor.
- * That is why again label is here: netlink_groups will be re-evaluated and
- * dump will be performed again.
- */
-	/* netlink_groups = 0; */
-	netlink_groups = RTMGRP_IPV4_IFADDR | RTMGRP_IPV6_IFADDR;
-
-	/* Already initialized? */
-	if (ie->nl.fd >= 0)
-		goto skip;
-	/* Prepare netlink. */
-	ul_nl_init(&(ie->nl));
-	if ((ul_netaddrq_init(&(ie->nl), NULL, NULL, (void *)ie)))
-		goto skip;
-
-	/* Open netlink and create address list. */
-	if (ul_nl_open(&(ie->nl),
-		       RTMGRP_LINK | netlink_groups))
-		goto skip;
-	if (ul_nl_request_dump(&(ie->nl), RTM_GETADDR))
-		goto error;
-	if (ul_nl_process(&(ie->nl), UL_NL_SYNC, UL_NL_LOOP) != UL_NL_DONE)
-		goto error;
-	goto skip;
-error:
-	/* In case of any error, the addrq list is just empty, and we can use
-	 * the code without any error checking. */
-	ul_nl_close(&(ie->nl));
-	ie->nl.fd = -1;
-skip:
-#endif /* USE_NETLINK */
-
 	free(ie->mem);
 	ie->mem = NULL;
 	ie->mem_sz = 0;
@@ -544,262 +505,6 @@ void agetty_issue_show(struct agetty_options *op)
 }
 
 #endif /* ISSUE_SUPPORT */
-
-#ifdef USE_NETLINK
-static void print_iface_best(struct agetty_issue *ie,
-			     const char *ifname,
-			     uint8_t ifa_family)
-{
-	struct ul_netaddrq_ip *best[__ULNETLINK_RATING_MAX];
-	struct ul_netaddrq_iface *ifaceq;
-	struct list_head *l;
-	enum ul_netaddrq_ip_rating threshold;
-
-	if (!ie->nl.data_addr)
-		return; /* error: init failed */
-
-	if ((ifaceq = ul_netaddrq_iface_by_name(&(ie->nl), ifname)))
-	{
-		memset(best, 0, sizeof(best));
-		if (ifa_family == AF_INET)
-			l = &(ifaceq->ip_quality_list_4);
-		else
-		/* if (ifa_family == AF_INET6) */
-			l = &(ifaceq->ip_quality_list_6);
-
-		threshold =
-			ul_netaddrq_iface_bestaddr(l, &best);
-		if (threshold != __ULNETLINK_RATING_MAX)
-			fputs(ul_nl_addr_ntop_address(best[threshold]->addr),
-			      ie->output);
-	}
-}
-
-static void print_addrq_bestofall(struct agetty_issue *ie,
-				  uint8_t ifa_family)
-{
-	struct ul_netaddrq_iface *best_ifaceq;
-	enum ul_netaddrq_ip_rating threshold;
-	const char *best_ipp;
-
-	if (!ie->nl.data_addr)
-		return; /* error: init failed */
-
-	best_ipp = ul_netaddrq_get_best_ipp(&(ie->nl), ifa_family,
-					    &threshold, &best_ifaceq);
-	if (best_ipp)
-		fputs(best_ipp, ie->output);
-}
-
-static void dump_iface_good(struct agetty_issue *ie,
-			    struct ul_netaddrq_iface *ifaceq)
-{
-	struct ul_netaddrq_ip *best4[__ULNETLINK_RATING_MAX];
-	struct ul_netaddrq_ip *best6[__ULNETLINK_RATING_MAX];
-	struct list_head *li;
-	enum ul_netaddrq_ip_rating threshold = __ULNETLINK_RATING_MAX - 1;
-	enum ul_netaddrq_ip_rating fthreshold; /* per family threshold */
-	bool first = true;
-
-	memset(best4, 0, sizeof(best4));
-	threshold = ul_netaddrq_iface_bestaddr(&(ifaceq->ip_quality_list_4),
-					       &best4);
-	memset(best6, 0, sizeof(best6));
-	fthreshold = ul_netaddrq_iface_bestaddr(&(ifaceq->ip_quality_list_6),
-						&best6);
-	if (fthreshold < threshold)
-		threshold = fthreshold;
-
-	list_for_each(li, &(ifaceq->ip_quality_list_4))
-	{
-		struct ul_netaddrq_ip *ipq;
-
-		ipq = list_entry(li, struct ul_netaddrq_ip, entry);
-		if (threshold <= ULNETLINK_RATING_SCOPE_LINK &&
-		    ( ipq->quality <= threshold ||
-		      /* Consider site addresses equally good as global */
-		      ipq->quality == ULNETLINK_RATING_SCOPE_SITE) &&
-		    best4[threshold])
-		{
-			if (first)
-			{
-				fprintf(ie->output, "%s: ", ifaceq->ifname);
-				first = false;
-			}
-			else
-				fprintf(ie->output, " ");
-			/* Write only the longest living temporary address */
-			if (threshold == ULNETLINK_RATING_F_TEMPORARY)
-			{
-				fputs(ul_nl_addr_ntop_address(best4[ULNETLINK_RATING_F_TEMPORARY]->addr),
-				      ie->output);
-				goto temp_cont4;
-			}
-			else
-				fputs(ul_nl_addr_ntop_address(ipq->addr),
-				      ie->output);
-		}
-	temp_cont4:;
-	}
-
-	list_for_each(li, &(ifaceq->ip_quality_list_6))
-	{
-		struct ul_netaddrq_ip *ipq;
-
-		ipq = list_entry(li, struct ul_netaddrq_ip, entry);
-		if (threshold <= ULNETLINK_RATING_SCOPE_LINK &&
-		    ( ipq->quality <= threshold ||
-		      /* Consider site addresses equally good as global */
-		      ipq->quality == ULNETLINK_RATING_SCOPE_SITE) &&
-		    best6[threshold])
-		{
-			if (first)
-			{
-				fprintf(ie->output, "%s: ", ifaceq->ifname);
-				first = false;
-			}
-			else
-				fprintf(ie->output, " ");
-			/* Write only the longest living temporary address */
-			if (threshold == ULNETLINK_RATING_F_TEMPORARY)
-			{
-				fputs(ul_nl_addr_ntop_address(best6[ULNETLINK_RATING_F_TEMPORARY]->addr),
-				      ie->output);
-				goto temp_cont6;
-			}
-			else
-				fputs(ul_nl_addr_ntop_address(ipq->addr),
-				      ie->output);
-		}
-	temp_cont6:;
-	}
-	if (!first)
-		fputs("\n", ie->output);
-}
-
-static void dump_iface_all(struct agetty_issue *ie,
-			   struct ul_netaddrq_iface *ifaceq)
-{
-	struct list_head *li;
-	struct ul_netaddrq_ip *ipq;
-	bool first = true;
-
-	list_for_each(li, &(ifaceq->ip_quality_list_4))
-	{
-		ipq = list_entry(li, struct ul_netaddrq_ip, entry);
-		if (first)
-		{
-			fprintf(ie->output, "%s: ", ifaceq->ifname);
-			first = false;
-		}
-		else
-			fprintf(ie->output, " ");
-		fputs(ul_nl_addr_ntop_address(ipq->addr), ie->output);
-	}
-	list_for_each(li, &(ifaceq->ip_quality_list_6))
-	{
-		ipq = list_entry(li, struct ul_netaddrq_ip, entry);
-		if (first)
-		{
-			fprintf(ie->output, "%s: ", ifaceq->ifname);
-			first = false;
-		}
-		else
-			fprintf(ie->output, " ");
-		fputs(ul_nl_addr_ntop_address(ipq->addr), ie->output);
-	}
-	if (!first)
-		fputs("\n", ie->output);
-}
-#endif /* USE_NETLINK */
-
-/*
- * parses \x{argument}, if not argument specified then returns NULL, the @fd
- * has to point to one char after the sequence (it means '{').
- */
-static char *get_escape_argument(FILE *fd, char *buf, size_t bufsz)
-{
-	size_t i = 0;
-	int c = fgetc(fd);
-
-	if (c == EOF || (unsigned char) c != '{') {
-		ungetc(c, fd);
-		return NULL;
-	}
-
-	do {
-		c = fgetc(fd);
-		if (c == EOF)
-			return NULL;
-		if ((unsigned char) c != '}' && i < bufsz - 1)
-			buf[i++] = (unsigned char) c;
-
-	} while ((unsigned char) c != '}');
-
-	buf[i] = '\0';
-	return buf;
-}
-
-static void output_special_char(struct agetty_issue *ie,
-				unsigned char c,
-				struct agetty_options *op,
-				struct termios *tp,
-				FILE *fp)
-{
-	struct utsname uts;
-
-	switch (c) {
-#ifdef USE_NETLINK
-	case '4':
-	case '6':
-	{
-		char iface[IF_NAMESIZE];
-		uint8_t ifa_family = c == '4' ? AF_INET : AF_INET6;
-
-		if (get_escape_argument(fp, iface, sizeof(iface)))
-			print_iface_best(ie, iface, ifa_family);
-		else
-			print_addrq_bestofall(ie, ifa_family);
-
-		/* TODO: Move to pass 1 */
-		if (c == '4')
-			netlink_groups |= RTMGRP_IPV4_IFADDR;
-		else
-			netlink_groups |= RTMGRP_IPV6_IFADDR;
-		break;
-	}
-	case 'a':
-	{
-		struct list_head *li;
-		struct ul_netaddrq_iface *ifaceq;
-
-		list_for_each_netaddrq_iface(li, &(ie->nl))
-		{
-			ifaceq = list_entry(li, struct ul_netaddrq_iface, entry);
-
-			dump_iface_good(ie, ifaceq);
-		}
-	}
-	break;
-	case 'A':
-	{
-		struct list_head *li;
-		struct ul_netaddrq_iface *ifaceq;
-
-		list_for_each_netaddrq_iface(li, &(ie->nl))
-		{
-			ifaceq = list_entry(li, struct ul_netaddrq_iface, entry);
-
-			dump_iface_all(ie, ifaceq);
-		}
-	}
-	break;
-#endif /* USE_NETLINK */
-	default:
-		putc(c, ie->output);
-		break;
-	}
-}
 
 
 void agetty_reload(void)
